@stephankramer 2023-11-02
Comparing timings [after](https://github.com/g-adopt/g-adopt/actions/runs/6728474362) and [latest master](https://github.com/g-adopt/g-adopt/actions/runs/6681954298) this seems to significantly speed up the first two tests on the list:

Before:

2d_cylindrical.py | 3101.448486s | Yes
-- | -- | --
3d_cartesian.py | 1731.114624s | Yes

After:

2d_cylindrical.py | 1363.541138s | Yes
-- | -- | --
3d_cartesian.py | 1288.952393s | Yes

but no significant changes for other tests (not sure why these two tbh).